requirements.txt
Django==2.2.6
django-ckeditor==5.8.0
django-js-asset==1.2.2
Pillow==6.2.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pytz==2019.2
sqlparse==0.3.0
